# docs/swagger.yaml
basePath: /
definitions:
...
  controllers.CreateProductRequest:
    properties:
      barcode:
        type: string
      image:
        type: string
      location:
        type: string
      name:
        type: string
      sku:
        type: string
      variant:
        type: string
    required:
    - barcode
    - image
    - location
    - name
    - sku
    - variant
    type: object
...
  controllers.ProductsListResponse:
    properties:
      pagination:
        $ref: '#/definitions/utils.PaginationResponse'
      products:
        items:
          $ref: '#/definitions/models.ProductResponse'
        type: array
    type: object
...
  controllers.UpdateProductRequest:
    properties:
      barcode:
        type: string
      image:
        type: string
      location:
        type: string
      name:
        type: string
      variant:
        type: string
    required:
    - barcode
    - image
    - location
    - name
    - variant
    type: object
...
  models.ProductResponse:
    properties:
      barcode:
        type: string
      created_at:
        type: string
      id:
        type: integer
      image:
        type: string
      location:
        type: string
      name:
        type: string
      sku:
        type: string
      updated_at:
        type: string
      variant:
        type: string
    type: object
...
  utils.PaginationResponse:
    properties:
      limit:
        type: integer
      page:
        type: integer
      total:
        type: integer
    type: object
  utils.Response:
    properties:
      data: {}
      error:
        type: string
      message:
        type: string
      success:
        type: boolean
    type: object
host: 192.168.31.136:8081
info:
  contact:
    email: support@example.com
    name: API Support
    url: http://www.example.com/support
  description: A comprehensive user management backend service with JWT authentication
    and role-based access control
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: Livotech Backend Service API
  version: "1.0"
paths:
...
  /api/products:
    get:
      consumes:
      - application/json
      description: Get list of all products with optional search by SKU (logged-in
        users only)
      parameters:
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 10
        description: Items per page
        in: query
        name: limit
        type: integer
      - description: Search by SKU (partial match)
        in: query
        name: search
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                data:
                  $ref: '#/definitions/controllers.ProductsListResponse'
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/utils.Response'
      security:
      - BearerAuth: []
      summary: Get all products
      tags:
      - products
    post:
      consumes:
      - application/json
      description: Create a new product (admin only)
      parameters:
      - description: Create product request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/controllers.CreateProductRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                data:
                  $ref: '#/definitions/models.ProductResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/utils.Response'
      security:
      - BearerAuth: []
      summary: Create new product
      tags:
      - products
  /api/products/{id}:
    delete:
      consumes:
      - application/json
      description: Soft delete a product (admin only)
      parameters:
      - description: Product ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/utils.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utils.Response'
      security:
      - BearerAuth: []
      summary: Remove product
      tags:
      - products
    get:
      consumes:
      - application/json
      description: Get specific product information (logged-in users only)
      parameters:
      - description: Product ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                data:
                  $ref: '#/definitions/models.ProductResponse'
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/utils.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utils.Response'
      security:
      - BearerAuth: []
      summary: Get product by ID
      tags:
      - products
    put:
      consumes:
      - application/json
      description: Update product information (admin only)
      parameters:
      - description: Product ID
        in: path
        name: id
        required: true
        type: integer
      - description: Update product request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/controllers.UpdateProductRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                data:
                  $ref: '#/definitions/models.ProductResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/utils.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utils.Response'
      security:
      - BearerAuth: []
      summary: Update product
      tags:
      - products
...
securityDefinitions:
  BearerAuth:
    description: Enter the token with the 'Bearer ' prefix, e.g. 'Bearer abcdef12345'.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"